requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
arch==7.2.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
attrs==26.1.0
bcrypt==5.0.0
blinker==1.9.0
blis==1.3.3
Bottleneck==1.6.0
build==1.4.3
cachetools==7.0.5
catalogue==2.0.10
catboost==1.2.8
certifi==2025.4.26
charset-normalizer==3.4.7
chromadb==1.5.7
click==8.3.2
click-plugins==1.1.1.2
cligj==0.7.2
cloudpathlib==0.23.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
confection==1.3.3
contourpy==1.3.3
curl_cffi==0.11.1
cycler==0.12.1
cymem==2.0.13
Cython==3.1.1
dacite==1.9.2
dataclasses-json==0.6.7
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1769745017/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
Deprecated==1.2.14
distro==1.9.0
durationpy==0.10
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1763918002538/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
filelock==3.28.0
fiona==1.10.1
flatbuffers==25.12.19
fonttools==4.62.1
frozendict==2.4.6
fsspec==2026.3.0
geopandas==1.1.2
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.74.0
graphviz==0.20.3
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
holiday==1.0.0
holidays==0.89
htmlmin==0.1.12
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.1
huggingface_hub==1.10.2
idna==3.11
ImageHash==4.3.1
img2pdf==0.5.1
importlib_metadata==8.7.1
importlib_resources==7.1.0
ipykernel @ file:///D:/bld/bld/rattler-build_ipykernel_1770566174/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1748711254/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_client_1767954655/work
jupyter_core @ file:///D:/bld/bld/rattler-build_jupyter_core_1760643889/work
kiwisolver==1.5.0
kubernetes==35.0.0
langchain==0.3.27
langchain-chroma==1.1.0
langchain-community==0.3.29
langchain-core==1.2.31
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.4.28
lightgbm==4.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.8
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multimethod==1.12
multitasking==0.0.11
murmurhash==1.0.15
narwhals==2.19.0
neo4j==5.28.2
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
numpy==2.4.4
oauthlib==3.3.1
onnxruntime==1.24.4
openai==2.31.0
openpyxl==3.1.5
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orjson==3.11.3
osmnx==2.0.7
overrides==7.7.0
packaging==25.0
pandas==2.3.2
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1770676243/work
peewee==3.18.1
phik==0.12.4
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pikepdf==8.13.0
pillow==10.2.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1775741140/work
plotly==6.7.0
pmdarima==2.0.4
preshed==3.0.13
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
protobuf==6.31.0
psutil @ file:///D:/bld/bld/rattler-build_psutil_1769678166/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
puremagic==1.29
pyarrow==23.0.1
pybase64==1.4.3
pydantic==2.11.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1774796700044/work
PyMuPDF==1.23.22
PyMuPDFb==1.23.22
pyogrio==0.12.1
pyparsing==3.3.2
PyPika==0.51.1
pyproj==3.7.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyvis==0.3.2
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487145/work
PyYAML==6.0.3
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1771716977/work
RapidFuzz==3.14.5
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
shap==0.47.2
shapely==2.1.2
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
slicer==0.0.8
smart_open==7.6.0
smmap==5.0.3
sniffio==1.3.1
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
squarify==0.4.4
srsly==2.5.3
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.13
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tornado @ file:///D:/bld/tornado_1774357957486/work
tqdm==4.67.3
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typeguard==4.4.2
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
visions==0.8.1
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1770634525285/work
weasel==1.0.0
websocket-client==1.9.0
websockets==15.0.1
wordcloud==1.9.4
wrapt==2.1.2
xgboost==3.0.2
xlsxwriter==3.2.9
ydata-profiling==4.16.1
yfinance==0.2.61
zipp==3.23.1
zstandard==0.25.0
